Final Version:
HOMEnVentory is an inventory management web application consisting of JSP's, Servlets, Beans and DB. This application consists of a Registration JSP, Login JSP, Admin JSP, Inventory JSP and Categories JSP, all run by servlets based on their respective names. 
The Login page verifies the user based on their existance, status and admin role. Inactive accounts will be shown the message of their account being deactivated and offering to register. 
If the role of the user is that of an Admin, the user will be send to the Admin page. In the Admin Page, there is a table of users on the left side, in this table you can see other registered Usernames, Passwords, Email, First and Last names, active status and Admin role, you can edit or delete each user. 
On the right side, there is a form for creating an account, what differentiates this form from the Registration form is that the Admin can use the status and admin role options when creating or editing. The Adding form also changes to the Edit User form where the User's information is filled for editing if a User is selected. Admins can also access the Categories page where they can Edit or Add a Category for the Inventory.
The Inventory Page is available to both users, in this page on the left side you have the Inventory table that displays your current inventory consisting of Category Name, Item Name and the Price, followed by Edit or Delete buttons. Underneath the table you can find the Add Item form that consists of Category, Item Name and Price. The form also changes into an Edit form if an Item is selected. 
On the right side of the page, a User table displays the current user's information and the user can then change their information if needed. If am Admin or User decides to logout, they will be sent to the Login page with a message showing successful logout.

This Application was created by Haseeb Sheikh.

Updates:

11/24/18
	- Created HOMEnVentory Application
	- Added login verification functionality.
	- Added Admin functionality such as:
		- Adding, editing and deleting users.
		- View functionality of users in a table.
	- Added Inventory management functionality such as:
		- Adding and deleting items.
		- View functionality of items in a table.

12/5/2018
	- Added registration page
	- Added Categories page
		- Can add a new Category.
		- Can edit a category.
	- Updated Admin functionality such as:
		- Adding, editing and deleting users.
		- View functionality of users in a table.
		- Update added active column to show status.
		- Update added active setting change.
	- Implemented security so normal users can't access admin features.
	- Implemented UI design via CSS.
	- Updated Inventory management functionality such as:
		- Adding and deleting items.
		- View functionality of items in a table.
		- Updated added edit function to items.

12/8/2018
	- Edited CSS format.
	- Fixed bugs and added security to admin delete functions.

12/9/2018
	- CSS changes to all pages.
	- Enhanced security for adding and registering a new user.